feat: Small margin between menu opener and menu
