+++ src/components/ActionMenuHelper/styles.styl
@@ -1,0 +1,3 @@
+.ActionMenuHelper__menu
+    position relative
+    top  .5rem
